Added WorkoutDetailActivity, Paused animation, minor database bug fixes:
- Added WorkoutDetailActivity
- Updated ComposeMapView to support dark theme
- Added Heat and HeatUnits for calculating calories
- Changed the Pause threshold value of WorkoutPauseDetector for better suiting normal environment.
- Added system bar padding to Launcher screen
- Added PhysicalUnitsPreference data class

+++ app/src/main/java/com/adityabavadekar/harmony/database/repo/WorkoutsRepository.kt
@@ -32,7 +32,7 @@ class WorkoutsRepository(
         return dao.getAllSummaryRecords()
     }
 
-    fun getWorkoutRecord(recordId: Int): WorkoutRecord? {
+    fun getWorkoutRecord(recordId: Long): WorkoutRecord? {
         return dao.getWorkoutRecord(recordId)
     }
